backend/requirements.txt
django==5.1.3
djangorestframework==3.15.2
django-cors-headers==4.4.0
requests==2.32.3
psycopg2-binary==2.9.10
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
drf-yasg==1.21.7
reportlab==4.2.2
whitenoise==6.7.0
